sending

+++ z25_run_hartree_lmirc.sh
@@ -99,7 +99,13 @@ elif [ ${status_build} == 0 ] ; then
         echo "Copied all log files to 9_all_lm_logs"
         echo
 
-        cp z_hartree-unsorted-${job_type}-${molecule_type}-${level_short}.csv ${dataset_results}/z_dataset-${molecule_type}-IRC-${naming_level}.csv
+
+
+        if [ ! -f ${dataset_results}/z_dataset-${molecule_type}-IRC-${naming_level}.csv} ] ; then
+            rm ${dataset_results}/z_dataset-${molecule_type}-IRC-${naming_level}.csv
+            cp z_hartree-unsorted-${job_type}-${molecule_type}-${level_short}.csv ${dataset_results}/z_dataset-${molecule_type}-IRC-${naming_level}.csv
+        fi
+
 
     fi
 fi
